New file added

+++ src/ex_140724/Lab003.java
@@ -9,7 +9,7 @@ public class Lab003 {
         //Numerical - byte, int, short, long, char
         //Floating - float, double
         // eg. we cant store a 128 in byte
-        byte a = 128;
+        byte a = (byte) 128;
         //Non Primitive Data Type - string, array, class, objects, enum
 
         //Constants
